Perturbation using standard_calc() without using work_list to control reference
Guessing now works. However as the number of pixels not tested when the first "glitched" pixel is found is sometimes nealrly all the pixels on the screen, finding the next useful reference point is very slow. The next reference may be even worse thn previous ones.

If we hit on a "good" reference, the speed is much greater using other plotting modes.